requirements.txt
python-dotenv==1.0.0
psycopg2-binary==2.9.9
langchain==0.2.14
openai==1.41.0
langchain_openai==0.1.21
pytz==2024.1
fastapi==0.115.0
uvicorn==0.31.0
alembic==1.13.3
SQLAlchemy==2.0.35
cloud-sql-python-connector==1.12.1
cloud-sql-python-connector[pg8000]==1.12.1
google-auth==2.35.0